Add persistence to bot
Dropped TinyDB from the import list, because it was unneeded.
Basically this will now create a db.json file on the local system
which will be used to keep persistence between python calls to the
bot. To illustrate how this is used i added examples for add/remove
in the whosin function and last shot winner in the shot-lottery function