Rework the event handler (part 1)
This is first commit to rework the event handler to make the code
more maintainable and extendable. The idea is that in order to add a
new command you make a new function for that command and add it to
the command dictionary. This prevents the handler from becoming too
large to debug and test.

This relates to number #9