Implement news feed
This will feed the latest twitter news from a variety of news sources
every couple of hours. The source is selected from the top of the
list and then put back in at the bottom. The order of the list is
shuffled each time the bot starts. More sources can be add using the
variable news_handles. All you need is the twitter handle.

This closes #3