Refactor get_last_tweet
This commit changes !trump back to the original behavior of printing
an error message on the twitter error. It does this by propagating
the error out of get_last_tweet. This implies as any caller of
get_last_tweet should handle TwythonError as appropriate.

This also fixes a minor error in get_news where the handler would not
move to the next source on failure. And removes an unnecessary global
definition in get_trump.