An auto-called shot lottery should be specific to the players playing
For now this is a simple work around that defaults to the largest
voice channel. So if !win calls a shot-lottery it will only include
players in the fullest voice channel. The also preserves the
!shot-lottery call to be all online members.

CSH indicated it might be preferable to have it only include people
that are in the channel from which the command is called. This is
entirely possible, and would indeed be preferable if we find that
groups are in multiple channels. For now, most seem to hang out in
the same channel and when using !win there is usually 3 or more
players in order for the win to go the record.

This closes #62