Initial push of poison Items
The next step here is make the death check a callable function so we
can call it at any point... Then to add the ability to self poison.

We will also have to think through how we want multiple different
poisons to work. Probably going with the highest damage, and still
adding to the duration.