Implement after action PUBG map reports
This branch/feature will create a map for a PUBG game that will show
truth traces and locations for deaths and kills. This will automagically
 be done for any games in the Top 10. But can be done on call for the
last match with !getmap.

Implements the !getmap command

This does not work in a container as the map generation requires a
connected UI (for tk). Future work will be around container detection
and handling code for disablement of this feature.